Rename the metatracer module to mt.
